Move deposit sweep code from `pkg/coordinator` to `pkg/maintainer/wallet`
Here we move all code related to deposit sweeps. By the way, some
simplifications were made in order to make the public API cleaner and
more readable.
